Add files via upload
This is a data analytics project created by using Power BI. Data from different countries are used in this project. The primary purpose of making this project is to analyze how the Zomato restaurant network works. It gives an idea in which country no. of restaurants available tomato with its ratings and online delivery added as a slicer.